DBZ-1591 Add parallel snapshot
Adds support for performing snapshots concurrently to the postgres and oracle
connectors. The snapshot process is now performed using a fixed worker thread
pool whose size is controlled by the "snapshot.max.threads" property.
Worker threads open a new connection to the datasource to snapshot a single
datasource object(ex. table) and close when finished. Connectors that do
support concurrent snapshots is a special case where the existing connection
is reused, performed on a fixed single thread pool, and all snapshotted objects
are published in order.